Add centos/redhat 7 support
The version scheme between RHEL7 and CentOS7 differ. This enburdens
us to make quarterly changes to keep this cookbook functioning if
we are explicit with versions.
By making systemd 'default' for rhel family, we can take advantage
of the regular version scheme of 6.x to correctly identify them
as 'sysv'. So, 'systemd' is now the default for this family.