Merge PR#224 - Fix runit container template
The current template tries to run two successive lines with exec. But since exec causes the current process to be replaced, the second line is never run. This causes runit to always report that the container service is down and to continually try to restart it.

Just remove the first exec to fix.