Merge PR#221 - Add NO_PROXY support
Thanks to jperville for this PR!
----
Even if little advertised, the docker daemon respects the NO_PROXY environment variable if present.

Setting this variable can be desirable to avoid passing through a proxy, for example when accessing a local docker registry (even more when the docker registry is running on an host that is not reachable from the http proxy such as localhost).

This PR adds a new attribute, node['docker']['no_proxy'] which will be injected in the default docker configuration, exactly like for the node['docker']['http_proxy'] attribute.