`parsed_checksum` helper method returns new_resource.checksum if specified
If a user wanted to use the `source` attribute for the `docker_service`
LWRP they need to specify a custom `checksum` attribute as well. The
`parsed_checksum` helper method was ignoring this custom attribute and
always providing the LWRP with a hard-coded checksum which matched with
the default docker binary source.