Change storage_opt attrib to storage_opts in docker_service LWRP
Currently if the `storage_opt` attribute for docker_service is
specified on the docker_service resource it's value will be passed
to the docker daemon prefixed by a single `--storage-opt` flag.

Each storage option passed to the docker daemon requires being
prefixed with it's own `--storage-opt` flag. Because of this
prefixing the value of `storage_opt` with a single `--storage-opt`
flag is a bug, as it makes passing multipe storage options to
the docker daemon hard to accomplish.

To correct this I've replaces the `storage_opt` attribute for
docker_service with `storage_opts` and accept an Array. Then in
the `docker_opts` helper method I collect the value of the array
and prefix each item with it's own `--storage-opt` flag.