Enable systemd options as a docker_service attribute.
Modifies the template file and systemd service files to allow users to specify an array of Systemd options,
such as "TasksMax=infinity" that need to go into the /etc/systemd/system/docker.service file.

Pull request #759 and associated issue #756 would be covered by this in that rather than assuming a default
for MountFlags that would need to apply to all Docker service deployments in Chef, users could dictate
these settings through an attribute/property of the docker_service resource or docker_service_manager
resource without losing the other extremely useful features of the cookbook by having to rewrite that file
with their own cookbooks.

It's not really right to assume that MountFlags will be the very last setting that developers will want
to change in the systemd settings--the list of potential settings is extremely long and complex.  Better
to let the developer make the decision in his/her cookbook.

Signed-off-by: Ben Smith <benjsmi@us.ibm.com>