By default setup the apt/yum repos in the package install resource
Having to fetch 2 cookbooks to use a resource is an anti-pattern. We have repository management built into chef now which we didn't have when this cookbook was originally written. The overhead of managing the repo automatically is pretty minimal. If someone doesn't want to do it they can disable it.

Signed-off-by: Tim Smith <tsmith@chef.io>