Require Chef Infra Client 15+ to fix issues with package versions
Chef Infra Client 15 reworked how versions are handled and made it so we can do versions like 19.03.3*. This means we can take the version in from the user and just do the right thing without constantly updating fragile helpers that build the full version string. We sadly still have that logic for Debian/Ubuntu, but baby steps.

Signed-off-by: Tim Smith <tsmith@chef.io>