Skip setting container swappiness with cgroupv2 (#1215)
* Skip setting container swappiness with cgroupv2

Setting swappiness with cgroupv2 and a CRI like crun will result in
containers that can be created but never started. Per-container
swappiness is not supported with cgroupv2.

* Use ubuntu/squid image for smoke suite

The ubuntu/squid image we had been using hasn't been updated in over 3 years and
this seems to be more maintained. Also set the ulimits to something higher so
that the tests pass properly.

Signed-off-by: Lance Albertson <lance@osuosl.org>

Signed-off-by: Lance Albertson <lance@osuosl.org>
Co-authored-by: Lance Albertson <lance@osuosl.org>